Starts a display module in ironrdp-dvc
I attempted to transfer everything in ironrdp_pdu::dvc::display
into it, but it required more refactoring that I'd expected to get it to
play nicely with the no_std possibility.

This also removes the requirement of keeping track of the channel_id
in each `DynamicVirtualChannel`, and instead lets the `DynamicChannelSet`
take care of returning that as needed.
