+++ crates/ironrdp-session/src/active_stage.rs
@@ -163,11 +163,6 @@ impl ActiveStage {
         Ok(vec![ActiveStageOutput::ResponseFrame(frame.into_inner())])
     }
 
-    // /// Sends a PDU on the dynamic channel.
-    // pub fn encode_dynamic(&self, output: &mut WriteBuf, channel_name: &str, dvc_data: &[u8]) -> SessionResult<()> {
-    //     self.x224_processor.encode_dynamic(output, channel_name, dvc_data)
-    // }
-
     /// Send a pdu on the static global channel. Typically used to send input events
     pub fn encode_static(&self, output: &mut WriteBuf, pdu: ShareDataPdu) -> SessionResult<usize> {
         self.x224_processor.encode_static(output, pdu)
